Expanded Logging. Changed Spreadsheet format. Refined Logs.
Added full logging capabilities, instead of only tracking a users last two messages or calls. Added the ability to track member joinings and leavings. Changed the formatting of the spreadsheet to accommodate these changes. Each user now has an individual sheet instead of a single row in an overall spreadsheet. Improved the print statements to plan for future bot logs in coming improvements.